backend/requirements.txt
django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
pytest-django==4.8.0
pytest-cov==4.1.0
python-dotenv==1.0.1
